use Int32 instead of int for team’s place
Int32 is enough to store a team’s place in a release and requires less memory.

+++ scripts/players.py
@@ -49,7 +49,7 @@ class PlayerRating:
 
     def update_places(self):
         self.data["place"] = (
-            self.data["rating"].rank(ascending=False, method="min").astype(int)
+            self.data["rating"].rank(ascending=False, method="min").astype("Int32")
         )
 
     def load_last_old_release(self):
